`vsc` module cleanup (#166)
A bunch of VSC-related renames per our chat. This way we won't conflict
between metrics reader and metrics provider

* Fixed FFI memory bug - `patience` was passing "off" without the null
termination to C
* Modified `patience`, refactoring it and making it return `Self` rather
than `Result<Self>`
* deleted `vsc` feature
* `mod vsc` -> inlined at the top crate level
* `Stats` -> `MetricsReader`
* `StatsImpl` -> `MetricsReaderImpl` (private)
* `StatsBuilder` -> `MetricsReaderBuilder`
* `Stat` -> `Metric`
* `Format` -> `MetricsFormat`
* Removed `Format::Unknown` value. This is a very `C-like` pattern
that's almost non-existent in Rust. Converted `From` to `TryFrom` -- as
it now can return an error - using `()` for now.

## TODO:
* [ ] Should `Semantics` be renamed into something?
* [ ] `patience` has an internal assert - is this needed? Should it
return a real result instead?
* [ ] `patience` uses `self.vsm`, but the `vsc_arg` below uses
`self.vsc`?
* [ ] The builder uses `Result<Self, Err>` return pattern - I'm not sure
how easy it will be to use... we might want to think if we want
`Result<&mut Self, Err>` instead? I'm not certain on this yet

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>