add stats macros (#163)
This adds macros for a vmod to expose stats (counters, gauges) into
varnish. For example:

```rust
#[varnish::stats]
pub struct Stats {
    // Number of hits
    #[counter]
    hits: std::sync::atomic::AtomicU64,

    // The temperature outside
    //
    // This tends to vary from one location to the next.
    #[gauge(level = "debug")]
    temperature: std::sync::atomic::AtomicU64,
}

let stats = Stats::new("mystats", "default");
stats.hits.fetch_add(1, std::sync::atomic::Ordering::Relaxed);
```

Then you would be able to see it in `varnishstat`:

```
$ varnishstat -1 |grep mystats
mystats.default.hits                       1         0.04 Number of hits
mystats.default.temperature                0         0.00 The temperature outside
```

## Implementation

This works by defining a struct to hold the counters. Varnish itself
will allocate the memory to hold the struct, and then in rust we cast
the memory to the struct type and fill it with values. This means we
need to be careful and only hold a reference to the struct rather than
copying it; I ended up with generating an inner struct from what the
user supplied and creating an outer struct with `Deref` to point to the
inner struct.

This currently generates the counter definition JSON by string
interpolation; if we're ok with `json_serde` as a dependency we could
use that instead.

---------

Co-authored-by: Yuri Astrakhan <YuriAstrakhan@gmail.com>